refactor: BusinessCardSetting -> contact 로 이동
